+++ dbt_platform_helper/utils/validation.py
@@ -409,6 +409,7 @@ ALB_DEFINITION = {
                 Optional("cdn_geo_restriction_type"): str,
                 Optional("cdn_logging_bucket"): str,
                 Optional("cdn_logging_bucket_prefix"): str,
+                Optional("cdn_timeout_seconds"): int,
                 Optional("default_waf"): str,
                 Optional("domain_prefix"): str,
                 Optional("enable_logging"): bool,